+++ sdk/storage/perf-tests/storage-blob/test/listBlobs.spec.ts
@@ -2,14 +2,14 @@
 // Licensed under the MIT license.
 
 import { generateUuid } from "@azure/core-http";
-import { PerfStressOptionDictionary, executeParallel } from "@azure/test-utils-perfstress";
+import { PerfOptionDictionary, executeParallel } from "@azure/test-utils-perf";
 import { StorageBlobTest } from "./storageTest.spec";
 interface StorageBlobListTestOptions {
   count: number;
 }
 
 export class StorageBlobListTest extends StorageBlobTest<StorageBlobListTestOptions> {
-  public options: PerfStressOptionDictionary<StorageBlobListTestOptions> = {
+  public options: PerfOptionDictionary<StorageBlobListTestOptions> = {
     count: {
       required: true,
       description: "Number of blobs to be listed",
@@ -29,7 +29,7 @@ export class StorageBlobListTest extends StorageBlobTest<StorageBlobListTestOpti
     );
   }
 
-  async runAsync(): Promise<void> {
+  async run(): Promise<void> {
     for await (const segmentResponse of this.containerClient.listBlobsFlat().byPage()) {
       // eslint-disable-next-line no-empty
       for (const _ of segmentResponse.segment.blobItems) {