+++ sdk/tables/perf-tests/data-tables/test/createComplexEntity.spec.ts
@@ -1,10 +1,10 @@
-import { PerfStressOptionDictionary } from "@azure/test-utils-perfstress";
+import { PerfOptionDictionary } from "@azure/test-utils-perf";
 import { TablesTest } from "./tables.spec";
 import { TableEntity } from "@azure/data-tables";
 import { createComplexEntity } from "./utils/createBaseEntity";
 
 export class CreateComplexEntityTest extends TablesTest {
-  public options: PerfStressOptionDictionary = {};
+  public options: PerfOptionDictionary = {};
 
   constructor() {
     super("ComplexEntityPerf");
@@ -18,7 +18,7 @@ export class CreateComplexEntityTest extends TablesTest {
     await super.globalCleanup();
   }
 
-  async runAsync(): Promise<void> {
+  async run(): Promise<void> {
     const complexEntity: TableEntity = createComplexEntity();
     await this.client.createEntity(complexEntity);
   }