+++ sdk/test-utils/perf/src/tests.ts
@@ -4,10 +4,10 @@
 import { AbortSignalLike } from "@azure/abort-controller";
 import { default as minimist, ParsedArgs as MinimistParsedArgs } from "minimist";
 import {
-  PerfStressOptionDictionary,
-  parsePerfStressOption,
-  DefaultPerfStressOptions,
-  defaultPerfStressOptions
+  PerfOptionDictionary,
+  parsePerfOption,
+  DefaultPerfOptions,
+  defaultPerfOptions
 } from "./options";
 import {
   TestProxyHttpClient,
@@ -18,33 +18,33 @@ import { HttpClient } from "@azure/core-http";
 import { Pipeline } from "@azure/core-rest-pipeline";
 
 /**
- * Defines the behavior of the PerfStressTest constructor, to use the class as a value.
+ * Defines the behavior of the PerfTest constructor, to use the class as a value.
  */
-export interface PerfStressTestConstructor<TOptions extends {} = {}> {
-  new (): PerfStressTest<TOptions>;
+export interface PerfTestConstructor<TOptions extends {} = {}> {
+  new (): PerfTest<TOptions>;
 }
 
 /**
- * Conveys the structure of any PerfStress test.
+ * Conveys the structure of any Perf test.
  * It allows developers to define the optional parameters to receive
  * in a way that aids them to auto-complete these options while adding code inside of each method.
  * It provides methods to parse the assigned options, as well as to set up and tear down any resources
- * at a global level (only one setup and teardown for the whole execution of a PerfStress test),
+ * at a global level (only one setup and teardown for the whole execution of a Perf test),
  * and at a local level, which happens once for each initialization of the test class
  * (initializations are as many as the "parallel" command line parameter specifies).
  */
-export abstract class PerfStressTest<TOptions = {}> {
+export abstract class PerfTest<TOptions = {}> {
   private readonly testProxy!: string;
   public testProxyHttpClient!: TestProxyHttpClient;
   public testProxyHttpClientV1!: TestProxyHttpClientV1;
-  public abstract options: PerfStressOptionDictionary<TOptions>;
+  public abstract options: PerfOptionDictionary<TOptions>;
 
   private static globalParallelIndex: number = 0;
   protected readonly parallelIndex: number;
 
   public constructor() {
-    this.parallelIndex = PerfStressTest.globalParallelIndex;
-    PerfStressTest.globalParallelIndex++;
+    this.parallelIndex = PerfTest.globalParallelIndex;
+    PerfTest.globalParallelIndex++;
 
     const testProxies = this.parsedOptions["test-proxies"].value;
     if (testProxies) {
@@ -53,15 +53,15 @@ export abstract class PerfStressTest<TOptions = {}> {
     }
   }
 
-  public get parsedOptions(): PerfStressOptionDictionary<TOptions & DefaultPerfStressOptions> {
+  public get parsedOptions(): PerfOptionDictionary<TOptions & DefaultPerfOptions> {
     // This cast is needed because TS thinks
-    //   PerfStressOptionDictionary<TOptions & DefaultPerfStressOptions>
+    //   PerfOptionDictionary<TOptions & DefaultPerfOptions>
     //   is different from
-    //   PerfStressOptionDictionary<TOptions> & PerfStressOptionDictionary<DefaultPerfStressOptions>
-    return parsePerfStressOption({
+    //   PerfOptionDictionary<TOptions> & PerfOptionDictionary<DefaultPerfOptions>
+    return parsePerfOption({
       ...this.options,
-      ...defaultPerfStressOptions
-    }) as PerfStressOptionDictionary<TOptions & DefaultPerfStressOptions>;
+      ...defaultPerfOptions
+    }) as PerfOptionDictionary<TOptions & DefaultPerfOptions>;
   }
 
   // Before and after running a bunch of the same test.
@@ -71,7 +71,7 @@ export abstract class PerfStressTest<TOptions = {}> {
   public setup?(): void | Promise<void>;
   public cleanup?(): void | Promise<void>;
 
-  public async runAsync?(abortSignal?: AbortSignalLike): Promise<void>;
+  public async run?(abortSignal?: AbortSignalLike): Promise<void>;
 
   /**
    * configureClientOptionsCoreV1
@@ -120,12 +120,10 @@ export abstract class PerfStressTest<TOptions = {}> {
 
 /**
  * Picks a specific test case by comparing the first command line paramter to the names of the
- * given classes, all of which must extend PerfStressTest.
- * @param tests An array of classes that extend PerfStressTest
+ * given classes, all of which must extend PerfTest.
+ * @param tests An array of classes that extend PerfTest
  */
-export function selectPerfStressTest(
-  tests: PerfStressTestConstructor[]
-): PerfStressTestConstructor {
+export function selectPerfTest(tests: PerfTestConstructor[]): PerfTestConstructor {
   const testsNames: string[] = tests.map((test) => test.name);
   const minimistResult: MinimistParsedArgs = minimist(process.argv);
   const testName = minimistResult._[minimistResult._.length - 1];