+++ sdk/test-utils/perf/test/exception.spec.ts
@@ -1,17 +1,17 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { PerfStressTest } from "../src";
+import { PerfTest } from "../src";
 
 /**
- * Exception is designed to test the response speed of the PerfStress test framework
+ * Exception is designed to test the response speed of the Perf test framework
  * If the option "sync" is passed, errors will be thrown on every test call, where the test being called is simple function.
  * Otherwise, errors thrown on every test call, where the test being called is a function returning a promise (an asynchronous function).
  */
-export class Exception extends PerfStressTest {
+export class Exception extends PerfTest {
   public options = {};
 
-  async runAsync(): Promise<void> {
+  async run(): Promise<void> {
     try {
       throw new Error();
     } catch (e) {