+++ sdk/test-utils/perf/test/sleep.spec.ts
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { PerfStressTest } from "../src";
+import { PerfTest } from "../src";
 import { delay } from "@azure/core-http";
 
 // Used for verifying the perf framework correctly computes average throughput across parallel tests of different speed
-export class SleepTest extends PerfStressTest {
+export class SleepTest extends PerfTest {
   private static instanceCount: number = 0;
   private secondsPerOperation: number = 0;
 
@@ -21,7 +21,7 @@ export class SleepTest extends PerfStressTest {
     this.secondsPerOperation = Math.pow(2, SleepTest.instanceCount);
   }
 
-  async runAsync(): Promise<void> {
+  async run(): Promise<void> {
     await delay(this.secondsPerOperation * 1000);
   }
 }